+++ EONIS_IT34_2020/EONIS_IT34_2020/Models/Entities/KontigentKarata.cs
@@ -11,7 +11,7 @@ namespace EONIS_IT34_2020.Models.Entities
         public String NazivKarte { get; set; }
         public String Sektor { get; set; }
         public String Ulaz { get; set; }
-        public double Cena { get; set; }
+        public int Cena { get; set; }
         public int Kolicina { get; set; }
         public String Napomena { get; set; }
         public Guid Id_administrator { get; set; }
